Prevent native qthread sync vars from being removed while they're alive
qthreads uses a global hash table to associate a sync var to it's FEB
state/queues. By default it removes the addr from the hash table when the FEB
is full and there are no pending operations.

I believe this hash table manipulation is expensive so allow sync vars to be
kept in the hash table, and update the sync var shim to keep them alive until
the chapel level sync var is destroyed.